requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# NLP & ML (Python 3.13 compatible versions)
transformers>=4.36.0
torch>=2.6.0
sentence-transformers>=2.2.2
spacy>=3.7.2
hanlp>=2.1.0b54

# Vector & Search
faiss-cpu>=1.7.4
whoosh>=2.7.4

# Graph Database
neo4j>=5.14.0
py2neo>=2021.2.3

# Web Framework
fastapi>=0.108.0
uvicorn>=0.25.0
streamlit>=1.29.0
requests>=2.31.0

# Data Processing
pandas>=2.1.4
numpy>=1.26.2
scikit-learn>=1.3.2
networkx>=3.2.1

# Evaluation
rank-eval==0.1.1

# Logging & Config
loguru==0.7.2
pyyaml==6.0.1
